iris_service/requirements.txt
# ============================================
# SentinelAgent Iris Service - Python Dependencies
# ============================================

# Web Framework
flask==3.0.0
flask-socketio==5.3.0
flask-cors==4.0.0

# WebSocket support
eventlet==0.33.3
python-socketio==5.9.0

# Image Processing
opencv-python==4.8.1.78
pillow==10.1.0

# Numerical Computing
numpy==1.26.2
scipy==1.11.4

# HTTP Client
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Production WSGI Server
gunicorn==21.2.0
